// user-client/src/types/enum/userEnum.ts
export enum SignupType {
    USER = 'USER', //일반 회원
    PARTNER = 'PARTNER', //파트너
    ADMIN = 'ADMIN' //관리자
}

export enum Gender {
  MALE = 'MALE',
  FEMALE = 'FEMALE',
}

export enum SocialAction {
    LOGIN = 'LOGIN',
    SIGNUP = 'SIGNUP'
}

export enum DeviceType { // 접속 디바이스 종류
  WEB = 'WEB',
  ANDROID = 'ANDROID',
  IOS = 'IOS'
}