// user-client/src/types/enum/ethEnum.ts
export enum ToggleStatus {
    ON = 'ON', //할성화
    OFF = 'OFF', //비황성화
}

export enum SettingType {
    PROFILE = 'PROFILE',  //프로필 수정 
    PASSWORD = 'PASSWORD', // 패스워드 수정
    SETTING = 'SETTING', //설정
    PARTNER_PROFILE = 'PARTNER_PROFILE' // 파트너 프로필
}

export enum BannerPosition {
    UP = 'UP',
    DOWN = 'DOWN'
}